### 20260125 * update [**ACPICA source**](https://github.com/acpica/acpica) **tag** **20251212** to build **ACPICA64** branch * update to **toro C Library v0.9.10Build 803**

+++ source/common/dmrestag.c
@@ -826,7 +826,7 @@ AcpiGetTagPathname (
     }
 
     (void) AcpiNsBuildNormalizedPath (BufferNode, Pathname,
-        RequiredSize, FALSE);
+        (UINT32)RequiredSize, FALSE);
 
     /*
      * Create the full path to the resource and tag by:
